Stream-data-real-time/requirements.txt
adal==1.2.7
alembic==1.14.0
applicationinsights==0.11.10
argcomplete==3.5.2
attrs==24.3.0
azure-common==1.1.28
azure-core==1.32.0
azure-eventhub==5.13.0
azure-functions==1.21.3
azure-graphrbac==0.61.2
azure-identity==1.19.0
azure-mgmt-authorization==4.0.0
azure-mgmt-containerregistry==10.3.0
azure-mgmt-core==1.5.0
azure-mgmt-keyvault==10.3.1
azure-mgmt-network==28.1.0
azure-mgmt-resource==23.2.0
azure-mgmt-storage==21.2.1
azureml==0.2.7
azureml-automl-core==1.59.0
azureml-core==1.59.0
azureml-dataprep==5.1.6
azureml-dataprep-native==41.0.0
azureml-dataprep-rslex==2.22.5
azureml-dataset-runtime==1.59.0
azureml-pipeline==1.59.0
azureml-pipeline-core==1.59.0
azureml-pipeline-steps==1.59.0
azureml-sdk==1.59.0
azureml-telemetry==1.59.0
azureml-train-automl-client==1.59.0
azureml-train-core==1.59.0
azureml-train-restclients-hyperdrive==1.59.0
backports.tempfile==1.0
backports.weakref==1.0.post1
bcrypt==4.2.1
blinker==1.8.2
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==2.2.1
colorama==0.4.6
contextlib2==21.6.0
contourpy==1.1.1
cryptography==44.0.0
cycler==0.12.1
databricks-sdk==0.40.0
Deprecated==1.2.15
docker==7.1.0
Flask==3.0.3
fonttools==4.55.3
fusepy==3.0.1
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.37.0
graphene==3.4.3
graphql-core==3.2.5
graphql-relay==3.2.0
greenlet==3.1.1
humanfriendly==10.0
idna==3.10
importlib_metadata==8.2.0
importlib_resources==6.4.0
isodate==0.7.2
itsdangerous==2.2.0
jeepney==0.8.0
Jinja2==3.1.5
jmespath==1.0.1
joblib==1.4.2
jsonpickle==4.0.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.7
knack==0.12.0
Mako==1.3.8
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.7.5
mlflow==2.17.2
mlflow-skinny==2.17.2
msal==1.31.1
msal-extensions==1.2.0
msrest==0.7.1
msrestazure==0.6.4.post1
ndg-httpsclient==0.5.1
numpy==1.23.5
oauthlib==3.2.2
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
packaging==24.2
pandas==2.0.3
paramiko==3.5.0
pathspec==0.12.1
pillow==10.4.0
pkginfo==1.12.0
pkgutil_resolve_name==1.3.10
portalocker==2.10.1
protobuf==5.29.2
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
Pygments==2.18.0
PyJWT==2.9.0
PyNaCl==1.5.0
pyOpenSSL==24.3.0
pyparsing==3.1.4
pyreadline3==3.5.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.2
pywin32==308
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.20.1
rsa==4.9
scikit-learn==1.3.2
scipy==1.10.1
SecretStorage==3.3.3
six==1.17.0
smmap==5.0.1
SQLAlchemy==2.0.36
sqlparse==0.5.3
tabulate==0.9.0
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
waitress==3.0.0
Werkzeug==3.0.6
wrapt==1.17.0
xgboost==2.1.3
zipp==3.20.2
confluent-kafka